增加购物车功能

+++ db.cpp
@@ -7,7 +7,7 @@ void insertGoods() {
         Goods* g=goods_list[pnu];
         std::string g_desc=g->getDesc();
         int g_type=g->getType();
-        double g_price=g->getPrice();
+        double g_price=g->getOriginPrice();
         int64_t g_remain=g->getRemain();
         QString cmd=QString("insert into goods values(%1,'").arg(id)+QString::fromStdString(pnu.first)+QString("','")+QString::fromStdString(g_desc)+QString("',%1,%2,%3,%4)").arg(g_type).arg(g_price).arg(g_remain).arg(pnu.second);
         query.exec(cmd);
@@ -16,8 +16,8 @@ void insertGoods() {
 }
 
 void changeGoods(const std::string& g_name,uint64_t uid) {
-    query.exec(QString("update goods set origin_price = %1 where name = '").arg(goods_list[std::pair(g_name,uid)]->getPrice())+QString::fromStdString(g_name)+QString("'"));
-    query.exec(QString("update goods set remain = %1 where name = '").arg(goods_list[std::pair(g_name,uid)]->getRemain())+QString::fromStdString(g_name)+QString("'"));
+    query.exec(QString("update goods set origin_price = %1 where name = '").arg(goods_list[std::pair(g_name,uid)]->getPrice())+QString::fromStdString(g_name)+QString("' and belong_to = %1").arg(uid));
+    query.exec(QString("update goods set remain = %1 where name = '").arg(goods_list[std::pair(g_name,uid)]->getRemain())+QString::fromStdString(g_name)+QString("' and belong_to = %1").arg(uid));
 }
 
 void deleteGoods(const std::string& g_name,uint64_t uid) {
@@ -30,7 +30,7 @@ void deleteGoods(const std::string& g_name,uint64_t uid) {
             i++;
         }
     }
-    query.exec(QString("delete from goods where name = '")+QString::fromStdString(g_name)+QString("'"));
+    query.exec(QString("delete from goods where name = '")+QString::fromStdString(g_name)+QString("' and belong_to = %1").arg(uid));
     query.exec(QString("update goods set id = id - 1 where id > %1").arg(index));
 }
 
@@ -53,19 +53,19 @@ void loadGoods() {
         std::pair<std::string,uint64_t> pnu(g_name,g_belong);
         switch (g_type) {
             case 0:
-                g=new Food(g_price,g_name,g_desc,g_remain);
+                g=new Food(g_price,g_name,g_desc,g_remain,g_belong);
                 goods_list[pnu]=g;
                 break;
             case 1:
-                g=new Cloths(g_price,g_name,g_desc,g_remain);
+                g=new Cloths(g_price,g_name,g_desc,g_remain,g_belong);
                 goods_list[pnu]=g;
                 break;
             case 2:
-                g=new Book(g_price,g_name,g_desc,g_remain);
+                g=new Book(g_price,g_name,g_desc,g_remain,g_belong);
                 goods_list[pnu]=g;
                 break;
             case 3:
-                g=new Electronic(g_price,g_name,g_desc,g_remain);
+                g=new Electronic(g_price,g_name,g_desc,g_remain,g_belong);
                 goods_list[pnu]=g;
                 break;
             default:
